refactor: service repository layer architecture
