+++ internal/app/app.go
@@ -4,8 +4,6 @@ import (
 	"context"
 	"log"
 
-	"github.com/kulikovroman08/reviewlink-backend/internal/service"
-
 	"github.com/gin-gonic/gin"
 	"github.com/jackc/pgx/v5/pgxpool"
 
@@ -14,6 +12,9 @@ import (
 	repoPlace "github.com/kulikovroman08/reviewlink-backend/internal/repository/place"
 	repoReview "github.com/kulikovroman08/reviewlink-backend/internal/repository/review"
 	repoUser "github.com/kulikovroman08/reviewlink-backend/internal/repository/user"
+	svcPlace "github.com/kulikovroman08/reviewlink-backend/internal/service/place"
+	svcReview "github.com/kulikovroman08/reviewlink-backend/internal/service/review"
+	svcUser "github.com/kulikovroman08/reviewlink-backend/internal/service/user"
 )
 
 func InitApp(cfg *configs.Config) *gin.Engine {
@@ -26,9 +27,12 @@ func InitApp(cfg *configs.Config) *gin.Engine {
 	reviewRepo := repoReview.NewPostgresReviewRepository(dbpool)
 	placeRepo := repoPlace.NewPostgresPlaceRepository(dbpool)
 
-	svc := service.NewService(userRepo, reviewRepo, placeRepo)
+	userSvc := svcUser.NewUserService(userRepo)
+	placeSvc := svcPlace.NewPlaceService(placeRepo)
 
-	app := controller.NewApplication(svc)
+	reviewSvc := svcReview.NewReviewService(reviewRepo, userRepo, placeRepo)
+
+	app := controller.NewApplication(userSvc, placeSvc, reviewSvc)
 
 	return controller.SetupRouter(app)
 }
